Server files put in separate folder
